compose_ui: Move `position_inside_code_block`.
This commit moves the `position_inside_code_block` function from
`compose_ui.ts` to `compose_textarea.ts` to prevent circular
dependencies down the line.

This is done as a prep for #29626.
